+++ R/auth.R
@@ -10,7 +10,6 @@
 #' @param package The name of the package authenticating
 #' @param cache Where to store authentication tokens
 #' @param use_oob Whther to use OOB browserless authetication
-#' @param new_user Deprecated, not used
 #'
 #' @return an OAuth token object, specifically a
 #'   \code{\link[=Token-class]{Token2.0}}, invisibly
@@ -50,12 +49,7 @@ gar_auth <- function(token = NULL,
                      app = gar_oauth_app(),
                      cache = gargle::gargle_oauth_cache(),
                      use_oob = gargle::gargle_oob_default(),
-                     package = "googleAuthR",
-                     new_user = NULL) {
-  
-  if(!is.null(new_user)){
-    warning("Argument new_user is deprecated and will be removed next release.")
-  }
+                     package = "googleAuthR") {
   
   # file locations to read existing httr tokens (legacy compatibility)
   if(is.string(token) && is.readable(token)){