+++ R/service_creation.R
@@ -32,7 +32,7 @@ gar_service_provision <- function(accountId,
   if(email == ""){
     email <- NULL
   }
-  gar_auth(email = email)
+  gar_auth(email = email, cache = FALSE)
   created <- gar_service_create(accountId, projectId = projectId)
     
   gar_service_grant_roles(created$email,