+++ R/setup_wizard_auth.R
@@ -14,6 +14,70 @@ gar_setup_auth_check <- function(env_arg = "GCE_AUTH_FILE"){
   TRUE
 }
 
+#' Check for a client JSON
+#' 
+#' @param session_user 1 for user level, 2 for project level, leave \code{NULL} to be prompted
+#' @param client_json The environment argument to be used for client_id/secret
+#' 
+#' @export
+#' @family setup functions
+#' @return TRUE is client_id is ready, FALSE if it is not
+gar_setup_clientid <- function(session_user = NULL,
+                               client_json = "GAR_CLIENT_JSON"){
+  
+  currently_set <- Sys.getenv(client_json)
+  
+  if(currently_set != ""){
+    valid <- validate_json(currently_set)
+    if(valid){
+      cli_alert_info("Using Client ID via {client_json}={currently_set}")
+      cli_rule()
+      return(TRUE)
+    }
+    
+    cli_alert_warning("Invalid client ID found - reconfiguring...")
+    
+  }
+  
+  cli_alert_info("Could not find a OAuth 2.0 Client ID via {client_json}")
+  
+  client_id <- usethis::ui_yeah("Have you downloaded a Client ID file?",
+                                yes = "Yes", no = "No")
+  
+  if(!client_id){
+    cli_alert_warning("You must have a client ID file to proceed.")
+    cli_alert_info("Download via https://console.cloud.google.com/apis/credentials/oauthclient :")
+    cli_li(c("Other > Name > Create >",
+             "OAuth 2.0 Client IDs >",
+             "Click Download Arrow to the right >",
+             "Download to your computer"))
+    cli_rule()
+    cli_alert_info("Rerun this wizard once you have your Client ID file")
+    if(usethis::ui_yeah("Open up service credentials URL?")){
+      utils::browseURL("https://console.cloud.google.com/apis/credentials/oauthclient")
+    }
+    return(FALSE)
+  }
+  
+  ff <- usethis::ui_yeah("Select location of your client ID file:",
+                         yes = "Ready", no = "Cancel")
+  
+  if(!ff){
+    return(FALSE)
+  }
+  
+  json <- file.choose()
+  valid <- validate_json(json)
+  
+  if(valid){
+    gar_setup_edit_renviron(paste0(client_json,"=",json), 
+                            session_user = session_user)
+  }
+  # we always return that R needs to be restarted and this needs to be rerun
+  FALSE
+  
+}
+
 #' Create a service account for googleCloudRunner
 #'
 #' This will use your Google OAuth2 user to create a suitable service account
@@ -37,63 +101,21 @@ gar_setup_auth_key <- function(email = Sys.getenv("GARGLE_EMAIL"),
                            roles = NULL,
                            default_key = "googleauthr"){
   
-  if(is.null(session_user)){
-    session_user <- gar_setup_check_session()
-  }
+  session_user <- gar_setup_check_session(session_user)
   
-  if(Sys.getenv(client_json) == ""){
-    cli_alert_info("Could not find a OAuth 2.0 Client ID via {client_json}")
-    
-    client_id <- usethis::ui_yeah("Have you downloaded a Client ID file?",
-                                  yes = "Yes", no = "No")
-    
-    if(!client_id){
-      cli_alert_warning("You must have a client ID file to proceed.")
-      cli_alert_info("Download via https://console.cloud.google.com/apis/credentials/oauthclient :")
-      cli_li(c("Other > Name > Create >",
-               "OAuth 2.0 Client IDs >",
-               "Click Download Arrow to the right >",
-               "Download to your computer"))
-      cli_rule()
-      cli_alert_info("Rerun this wizard once you have your Client ID file")
-      if(usethis::ui_yeah("Open up service credentials URL?")){
-        utils::browseURL("https://console.cloud.google.com/apis/credentials/oauthclient")
-      }
-      return(FALSE)
-    }
-    
-    ff <- usethis::ui_yeah("Select location of your client ID file:",
-                           yes = "Ready", no = "Cancel")
-    
-    if(!ff){
-      return(FALSE)
-    }
-    
-    json <- file.choose()
-    valid <- validate_json(json)
-    
-    if(valid){
-      gar_setup_edit_renviron(paste0(client_json,"=",json), 
-                              session_user = session_user)
-    }
-    # we always return that gar_setup() needs to be rerun
-    return(FALSE)
-    
-  }
+  client_done <- gar_setup_clientid(session_user, client_json = client_json)
   
-  cli_alert_info("Using Client ID via {client_json}")
-  cli_rule()
-  json <- Sys.getenv(client_json)
-  if(!validate_json(json)){
+  if(!client_done){
+    cli_alert_danger("Need a clientId to be set before configuring further")
     return(FALSE)
   }
   
-  create_service <- usethis::ui_yeah("Client ID present but no service authentication file is configured.  Do you want to provision a service account for your project?",
+  create_service <- usethis::ui_yeah("Do you want to provision a service account for your project?",
                                      yes = "Yes, I need a service account key",
                                      no = "No, I already have one downloaded")
   
   if(!create_service){
-    cli_alert_danger("No service account provisioned, using existing")
+    cli_alert_info("No service account provisioned, using existing")
     return(TRUE)
   }
   
@@ -112,7 +134,7 @@ gar_setup_auth_key <- function(email = Sys.getenv("GARGLE_EMAIL"),
   gar_service_provision(
     account_id,
     roles = roles,
-    json = json,
+    json = Sys.getenv(client_json),
     file = file,
     email = email)
   
@@ -132,16 +154,16 @@ gar_setup_auth_key <- function(email = Sys.getenv("GARGLE_EMAIL"),
 
 #' Setup wizard helper - add authentication file to .Renviron
 #' 
-#' @param session_user 1 if its a session or 2 if user based .Renviron
+#' @param ... Other arguments passed to \link{gar_setup_auth_key}
 #' @param env_arg The environment argument to set
 #' 
 #' @return A string to paste into an .Renviron, or NULL
 #' @export
 #' @importFrom cli cli_alert_danger cli_alert_success
-gar_setup_get_authenv <- function(session_user, 
-                                  env_arg = "GCE_AUTH_FILE"){
+gar_setup_get_authenv <- function(env_arg = "GCE_AUTH_FILE", 
+                                  ...){
   # checks if client JSON present
-  auth <- gar_setup_auth_key(session_user = session_user)
+  auth <- gar_setup_auth_key(...)
   
   if(!auth){
     return(NULL)