+++ NEWS.md
@@ -3,7 +3,8 @@
 * Let you supply your own URL for discovery documents in `gar_discovery_api()`
 * Change defaults on `gar_create_package()` to check = FALSE, github=FALSE
 * Add class 'list' to `gar_discovery_*()` objects so they work with `jsonlite`
-* Add `gar_service_list`
+* Add `gar_service_list()` and `gar_service_get()`
+* If using `gar_service_create()` and it already exists, will return existing via `gar_service_get()`
 
 # googleAuthR 1.2.1