+++ NEWS.md
@@ -1,4 +1,4 @@
-# googleAuthR 1.2.1.9000
+# googleAuthR 1.3.0
 
 * Let you supply your own URL for discovery documents in `gar_discovery_api()`
 * Change defaults on `gar_create_package()` to check = FALSE, github=FALSE